Refactor to own component like a civilised developer
